Update rationale.html, add link to article for downloading
The page lamented the fact that the article "A Multiple Substring Search Algorithm" (Halibard / Rubin) was not available on the web.  In fact, it can be accessed from several sites, the nicest one (in color) being ResearchGate, which I have updated on the page.

There is another URL: http://www.mountainvistasoft.com/personal/articles/halibard-rubin-cuj-june-2002.pdf, which could be added as an additional cached copy on the web.